// Neptune.Web/src/app/shared/generated/model/treatment-bmpby-type-grid-dto.ts
export class TreatmentBMPByTypeGridDto { 
    TreatmentBMPID?: number;
    TreatmentBMPName?: string | null;
    StormwaterJurisdictionID?: number;
    StormwaterJurisdictionName?: string | null;
    OwnerOrganizationName?: string | null;
    YearBuilt?: number | null;
    SystemOfRecordID?: string | null;
    WaterQualityManagementPlanID?: number | null;
    WaterQualityManagementPlanName?: string | null;
    Notes?: string | null;
    LatestAssessmentDate?: string | null;
    LatestAssessmentScore?: number | null;
    NumberOfAssessments?: number;
    LatestMaintenanceDate?: string | null;
    NumberOfMaintenanceRecords?: number;
    BenchmarkAndThresholdSet?: boolean;
    TreatmentBMPLifespanTypeDisplayName?: string | null;
    TreatmentBMPLifespanEndDate?: string | null;
    RequiredFieldVisitsPerYear?: number | null;
    RequiredPostStormFieldVisitsPerYear?: number | null;
    SizingBasisTypeDisplayName?: string | null;
    TrashCaptureStatusTypeDisplayName?: string | null;
    DelineationTypeDisplayName?: string | null;
    CustomAttributeValues?: { [key: string]: string; } | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}